# openapi.yaml
openapi: 3.1.0
info:
  title: Books & Finance API
  version: 1.0.0
  description: |
    Personal finance and books management REST API.

    **Base URL:** `http://localhost:8000/api/v1`

    **Authentication:** Bearer JWT token. Obtain via `POST /auth/login`.
    Tokens expire in **7 days**.

    **Default credentials:** `BTC007` / `password123`
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    search:
      name: search
      in: query
      schema: { type: string }
    from:
      name: from
      in: query
      schema: { type: string, format: date }
    to:
      name: to
      in: query
      schema: { type: string, format: date }
security:
  - bearerAuth: []
paths:
...
  /expenses:
    get:
      summary: List expenses
      tags: [Expenses]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
      responses:
        '200': { description: Expenses list }
    post:
      summary: Add expense
      tags: [Expenses]
      requestBody:
        required: true
        content:
          application/json:
            example:
              category: Food
              amount: 45.00
              date: "2024-01-15"
              description: Dinner
      responses:
        '201': { description: Expense created }
  /stock:
    get:
      summary: List stock items
      tags: [Stock]
      parameters:
        - $ref: '#/components/parameters/search'
        - name: category
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
      responses:
        '200': { description: Stock list }
    post:
      summary: Create stock item
      tags: [Stock]
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: The Great Gatsby
              sku: BOOK-001
              quantity: 50
              unit_price: 12.99
              category: Fiction
              location: Shelf A3
      responses:
        '201': { description: Stock item created }
  /stock/{id}/adjust-quantity:
    patch:
      summary: Adjust stock quantity by delta
      tags: [Stock]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [delta]
              properties:
                delta:
                  type: integer
                  description: Positive to add, negative to subtract. Quantity cannot go below 0.
                  example: -5
      responses:
        '200': { description: Quantity adjusted }
        '400': { description: Would result in negative quantity }
...
  /split-expenses:
    get:
      summary: List split expenses
      tags: [Split Expenses]
      responses:
        '200': { description: Split expenses list }
    post:
      summary: Create split expense with participants
      tags: [Split Expenses]
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: Team Dinner
              total_amount: 120.00
              date: "2024-01-20"
              notes: Restaurant bill
              participants:
                - name: Alice
                  share_amount: 40.00
                - name: Bob
                  share_amount: 40.00
                - name: Charlie
                  share_amount: 40.00
      responses:
        '201': { description: Split expense created }
        '400':
          description: Participant shares do not sum to total_amount
  /split-expenses/{id}/participants/{participantId}/settle:
    patch:
      summary: Mark a participant as settled
      tags: [Split Expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: participantId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Participant settled }
...